# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAPhoto API
  description: |-
    Keep in touch with your friends by sharing photos of special moments, thanks to WASAPhoto! You can
    upload your photos directly from your PC, and they will be visible to everyone following you.
  version: "3.0"
...
paths:
...
  /users/{user_id}/photos:
    parameters:
      - name: user_id
        in: path
        schema:
          $ref: "#/components/schemas/uid"
        required: true
        description: The user ID of the user who uploads the photo.

    get:
      tags: ["photos"]
      summary: Returns user photos
      description: Returns the list of photos uploaded by a user.
      operationId: getUserPhotos
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            $ref: "#/components/schemas/limit"
          description: The number of elements to show.
          required: false
        - name: start_index
          in: query
          schema:
            $ref: "#/components/schemas/start_index"
          description: The starting offset.
          required: false
      responses:
        '200':
          description: Returns the user photos list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/user_photo_stream"
        '404':
          description: User not found (or the authorized user is banned).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generic_response"
              example:
                status: "Resource not found"
        '400':
          description: Some parameters are malformed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generic_response"
              example:
                status: "Invalid start_index or limit value"

    post:
      tags: ["photos"]
      summary: Uploads a photo
      description: Uploads a photo in the gallery of the authorized user.
      operationId: uploadPhoto
      security:
        - BearerAuth: []
      requestBody:
        content:
          image/jpeg:
            schema:
              description: The photo to upload.
              type: string
              format: binary
      responses:
        '201':
          description: Upload photo action successful.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generic_response"
              example:
                status: "Resource created"
        '404':
          description: User not found (or the authorized user is banned).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generic_response"
              example:
                status: "Resource not found"
  /users/{user_id}/photos/{photo_id}:
    parameters:
      - name: user_id
        in: path
        schema:
          $ref: "#/components/schemas/uid"
        description: The ID of the author of the photo.
        required: true
      - name: photo_id
        in: path
        schema:
          $ref: "#/components/schemas/photo_id"
        description: The ID of the photo.
        required: true
    get:
      tags: ["photos"]
      summary: Downloads a photo
      description: Returns the requested photo in binary format.
      operationId: getUserPhoto
      security:
        - BearerAuth: []
      responses:
        '200':
          description: The requested photo in binary format.
          content:
            image/jpeg:
              schema:
                #todo it wants lenght
                description: The requested photo in binary format.
                type: string
                format: binary
        '404':
          description: User or photo not found (or the authorized user is banned by the author of the photo).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generic_response"
              example:
                status: "Resource not found"
    delete:
      tags: ["photos"]
      summary: Deletes a photo
      description: Deletes a photo in the gallery of the authorized user.
      operationId: deletePhoto
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Delete photo action successful.
        '401':
          description: The user has no permission to delete that photo.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generic_response"
              example:
                status: "Unauthorized"
        '404':
          description: User or photo not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generic_response"
              example:
                status: "Resource not found"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    start_index:
      type: integer
      description: The starting offset.
      default: 0
    limit:
      type: integer
      description: The number of elements to show.
      default: 15
...
    uid:
      type: string
      minLength: 36
      maxLength: 36
      description: The user ID.
      format: uuid
      example: "1b4e28ba-2fa1-11d2-883f-0016d3cca427"
...
    photo_id:
      type: integer
      description: The ID of the photo.
      example: 1527
...
    upload_time:
      type: string
      format: date-time
      description: Upload time and date in RFC3339 format.
      minLength: 25
      maxLength: 25
    likes:
      type: integer
      example: 90
      description: Number of likes.
    comments_n:
      type: integer
      example: 7
      description: Number of comments on the photo.
...
    like_status:
      type: boolean
      example: true
      description: Whether the user liked the photo.
...
    user_photo_stream:
      type: array
      minItems: 0
      maxItems: 100
      description: An array of photos, upload time and likes.
      items:
        $ref: "#/components/schemas/user_photo_stream_item"
    user_photo_stream_item:
      type: object
      properties:
        photo_id:
          $ref: "#/components/schemas/photo_id"
        upload_time:
          $ref: "#/components/schemas/upload_time"
        likes:
          $ref: "#/components/schemas/likes"
        comments:
          $ref: "#/components/schemas/comments_n"
        liked:
          $ref: "#/components/schemas/like_status"
...
    generic_response:
      type: object
      description: An object representing a success.
      properties:
        status:
          type: string
          pattern: ".*"
          description: The status of the request.
          example: "Success"
...